better support for gfx_tileset_load_* by passing gfx_tileset_options to all
