+++ CMakeLists.txt
@@ -38,6 +38,7 @@ zvb_add_library(zvb_gfx
     ${GFX_DIR}/tilemap/load.c
     ${GFX_DIR}/tilemap/place.c
     ${GFX_DIR}/tileset/load.c
+    ${GFX_DIR}/tileset/load_none.c
     ${GFX_DIR}/tileset/load_1bit.c
     ${GFX_DIR}/tileset/load_2bit.c
     ${GFX_DIR}/tileset/load_4bit.c
@@ -67,6 +68,7 @@ zvb_add_library(zvb
     ${GFX_DIR}/tilemap/load.c
     ${GFX_DIR}/tilemap/place.c
     ${GFX_DIR}/tileset/load.c
+    ${GFX_DIR}/tileset/load_none.c
     ${GFX_DIR}/tileset/load_1bit.c
     ${GFX_DIR}/tileset/load_2bit.c
     ${GFX_DIR}/tileset/load_4bit.c
